hercule is dead, long live to hermes
